requirements.txt
fastapi
uvicorn
httpx
confluent_kafka
kafka-python
